refactor(ActionsCard): ♻️ moved some info to InfoCard
